Pass `dtype` as string to visualization components
